seperate out main image to develope zoom feature
